Simplify the abstraction for one-time & repetitive tasks
