+++ classes/suggested-tasks/providers/class-sample-page.php
@@ -5,15 +5,14 @@
  * @package Progress_Planner
  */
 
-namespace Progress_Planner\Suggested_Tasks\Providers\One_Time;
+namespace Progress_Planner\Suggested_Tasks\Providers;
 
-use Progress_Planner\Suggested_Tasks\Providers\One_Time;
 use Progress_Planner\Suggested_Tasks\Data_Collector\Sample_Page as Sample_Page_Data_Collector;
 
 /**
  * Add task to delete the Sample Page.
  */
-class Sample_Page extends One_Time {
+class Sample_Page extends Tasks {
 
 	/**
 	 * The provider ID.
