+++ classes/suggested-tasks/providers/class-tasks.php
@@ -20,7 +20,7 @@ abstract class Tasks implements Tasks_Interface {
 	 *
 	 * @var string
 	 */
-	protected const CATEGORY = '';
+	protected const CATEGORY = 'configuration';
 
 	/**
 	 * The ID of the task provider.
@@ -42,6 +42,13 @@ abstract class Tasks implements Tasks_Interface {
 	 * @var bool
 	 */
 	protected const IS_ONBOARDING_TASK = false;
+
+	/**
+	 * Whether the task is repetitive.
+	 *
+	 * @var bool
+	 */
+	protected $is_repetitive = false;
 
 	/**
 	 * The task points.
@@ -216,10 +223,25 @@ abstract class Tasks implements Tasks_Interface {
 	/**
 	 * Get the task ID.
 	 *
+	 * @param array $data Optional data to include in the task ID.
 	 * @return string
 	 */
-	public function get_task_id() {
-		return $this->get_provider_id();
+	public function get_task_id( $data = [] ) {
+		if ( ! $this->is_repetitive() ) {
+			return $this->get_provider_id();
+		}
+
+		$parts = [ $this->get_provider_id() ];
+
+		// Add optional data parts if provided.
+		if ( ! empty( $data['post_id'] ) ) {
+			$parts[] = $data['post_id'];
+		}
+
+		// Always add the date as the last part.
+		$parts[] = \gmdate( 'YW' );
+
+		return implode( '-', $parts );
 	}
 
 	/**
@@ -231,6 +253,15 @@ abstract class Tasks implements Tasks_Interface {
 		return static::CAPABILITY
 			? \current_user_can( static::CAPABILITY )
 			: true;
+	}
+
+	/**
+	 * Check if the task is a repetitive task.
+	 *
+	 * @return bool
+	 */
+	public function is_repetitive() {
+		return $this->is_repetitive;
 	}
 
 	/**
@@ -290,5 +321,159 @@ abstract class Tasks implements Tasks_Interface {
 	 */
 	public function is_task_relevant() {
 		return true;
+	}
+
+	/**
+	 * Evaluate a task.
+	 *
+	 * @param string $task_id The task ID.
+	 *
+	 * @return false|\Progress_Planner\Suggested_Tasks\Task The task data or false if the task is not completed.
+	 */
+	public function evaluate_task( $task_id ) {
+		// Early bail if the user does not have the capability to manage options.
+		if ( ! $this->capability_required() ) {
+			return false;
+		}
+
+		if ( ! $this->is_repetitive() ) {
+			if ( 0 !== strpos( $task_id, $this->get_task_id() ) ) {
+				return false;
+			}
+			return $this->is_task_completed() ? Task_Factory::create_task_from( 'id', $task_id ) : false;
+		}
+
+		$task_object = Task_Factory::create_task_from( 'id', $task_id );
+		$task_data   = $task_object->get_data();
+
+		if ( $task_data['provider_id'] === $this->get_provider_id() && \gmdate( 'YW' ) === $task_data['date'] && $this->is_task_completed( $task_id ) ) {
+			// Allow adding more data, for example in case of 'create-post' tasks we are adding the post_id.
+			$task_data = $this->modify_evaluated_task_data( $task_data );
+			$task_object->set_data( $task_data );
+
+			return $task_object;
+		}
+
+		return false;
+	}
+
+	/**
+	 * Check if the task condition is satisfied.
+	 * (bool) true means that the task condition is satisfied, meaning that we don't need to add the task or task was completed.
+	 *
+	 * @return bool
+	 */
+	abstract protected function should_add_task();
+
+	/**
+	 * Alias for should_add_task(), for better readability when using in the evaluate_task() method.
+	 *
+	 * @param string $task_id Optional task ID to check completion for.
+	 * @return bool
+	 */
+	public function is_task_completed( $task_id = '' ) {
+		// If no specific task ID provided, use the default behavior.
+		return ! $this->is_repetitive() || ! empty( $task_id )
+			? ! $this->should_add_task()
+			: $this->is_specific_task_completed( $task_id );
+	}
+
+	/**
+	 * Check if a specific task is completed.
+	 * Child classes can override this method to handle specific task IDs.
+	 *
+	 * @param string $task_id The task ID to check.
+	 * @return bool
+	 */
+	protected function is_specific_task_completed( $task_id ) {
+		return ! $this->should_add_task();
+	}
+
+	/**
+	 * Backwards-compatible method to check if the task condition is satisfied.
+	 *
+	 * @return bool
+	 */
+	protected function check_task_condition() {
+		return ! $this->should_add_task();
+	}
+
+	/**
+	 * Get an array of tasks to inject.
+	 *
+	 * @return array
+	 */
+	public function get_tasks_to_inject() {
+
+		$task_id = $this->get_task_id();
+
+		if (
+			true === $this->is_task_snoozed() ||
+			! $this->should_add_task() || // No need to add the task.
+			true === \progress_planner()->get_suggested_tasks()->was_task_completed( $this->get_task_id() )
+		) {
+			return [];
+		}
+
+		$task_data = [
+			'task_id'     => $task_id,
+			'provider_id' => $this->get_provider_id(),
+			'category'    => $this->get_provider_category(),
+			'date'        => \gmdate( 'YW' ),
+		];
+
+		$task_data = $this->modify_injection_task_data( $task_data );
+
+		return [
+			$task_data,
+		];
+	}
+
+	/**
+	 * Modify task data before injecting it.
+	 * Child classes can override this method to add extra data.
+	 *
+	 * @param array $task_data The task data.
+	 *
+	 * @return array
+	 */
+	protected function modify_injection_task_data( $task_data ) {
+		return $task_data;
+	}
+
+	/**
+	 * Modify task data after task was evaluated.
+	 * Child classes can override this method to add extra data.
+	 *
+	 * @param array $task_data The task data.
+	 *
+	 * @return array
+	 */
+	protected function modify_evaluated_task_data( $task_data ) {
+		return $task_data;
+	}
+
+	/**
+	 * Get the task details.
+	 *
+	 * @param string $task_id The task ID.
+	 *
+	 * @return array
+	 */
+	public function get_task_details( $task_id = '' ) {
+
+		return [
+			'task_id'      => $this->get_task_id(),
+			'provider_id'  => $this->get_provider_id(),
+			'title'        => $this->get_title(),
+			'parent'       => $this->get_parent(),
+			'priority'     => $this->get_priority(),
+			'category'     => $this->get_provider_category(),
+			'points'       => $this->get_points(),
+			'url'          => $this->capability_required() ? \esc_url( $this->get_url() ) : '',
+			'description'  => $this->get_description(),
+			'link_setting' => $this->get_link_setting(),
+			'dismissable'  => $this->is_dismissable(),
+		];
 	}
 }
